Change to profile to prevent auto cropping. App now checks if the Completed folder exists and tries to create it if it doesn't.
git-svn-id: svn://webhost01/Personal%20Projects@263 ebe4f960-59b0-3b4e-9416-0a66d08d054f

+++ HandbrakeBatchEncoder/HandbrakeEncoder.cs
@@ -93,7 +93,21 @@ namespace HandbrakeBatchEncoder
                 string moveToPath = Path.Combine(Settings.Default.WatchFolder, Settings.Default.CompletedInputFolder);
 
                 // Move the source file to a Completed subfolder
-                File.Move(_sourceFile, Path.Combine(moveToPath, Path.GetFileName(_sourceFile)));
+                if (!Directory.Exists(moveToPath))
+                {
+                    // Folder doesn't exist - try to create it
+                    try
+                    {
+                        Directory.CreateDirectory(moveToPath);
+                    }
+                    catch (IOException)
+                    {}
+                }
+
+                if (Directory.Exists(moveToPath))
+                {
+                    File.Move(_sourceFile, Path.Combine(moveToPath, Path.GetFileName(_sourceFile)));
+                }
 
             }
